+++ ydb/_topic_writer/topic_writer_asyncio.py
@@ -478,7 +478,12 @@ class WriterAsyncIOReconnector:
                 tasks = [send_loop, receive_loop]
                 done, _ = await asyncio.wait([send_loop, receive_loop], return_when=asyncio.FIRST_COMPLETED)
                 done.pop().result()  # need for raise exception - reason of stop task
-            except issues.Error as err:
+            except (asyncio.CancelledError, issues.Error) as err:
+                if isinstance(err, asyncio.CancelledError):
+                    if self._closed:
+                        return
+                    err = issues.ConnectionLost("gRPC stream cancelled")
+
                 err_info = check_retriable_error(err, retry_settings, attempt)
                 if not err_info.is_retriable or self._tx is not None:  # no retries in tx writer
                     logger.debug("writer reconnector %s stop connection loop due to %s", self._id, err)
@@ -492,7 +497,7 @@ class WriterAsyncIOReconnector:
                 )
                 await asyncio.sleep(err_info.sleep_timeout_seconds)
 
-            except (asyncio.CancelledError, Exception) as err:
+            except Exception as err:
                 self._stop(err)
                 return
             finally:
@@ -511,13 +516,20 @@ class WriterAsyncIOReconnector:
                     messages.extend(self._messages_for_encode.get_nowait())
 
                 logger.debug(
-                    "writer reconnector %s encode %s messages",
+                    "writer reconnector %s start encoding %s messages",
                     self._id,
                     len(messages),
                 )
 
                 batch_codec = await self._codec_selector(messages)
                 await self._encode_data_inplace(batch_codec, messages)
+
+                logger.debug(
+                    "writer reconnector %s encoded %s messages",
+                    self._id,
+                    len(messages),
+                )
+
                 self._add_messages_to_send_queue(messages)
         except BaseException as err:
             self._stop(err)
@@ -626,6 +638,8 @@ class WriterAsyncIOReconnector:
             for ack in resp.acks:
                 self._handle_receive_ack(ack)
 
+            logger.debug("writer reconnector %s handled %s acks", self._id, len(resp.acks))
+
     def _handle_receive_ack(self, ack):
         current_message = self._messages.popleft()
         message_future = self._messages_future.popleft()
@@ -645,12 +659,6 @@ class WriterAsyncIOReconnector:
         else:
             raise TopicWriterError("internal error - receive unexpected ack message.")
         message_future.set_result(result)
-        logger.debug(
-            "writer reconnector %s ack seqno=%s result=%s",
-            self._id,
-            ack.seq_no,
-            type(result).__name__,
-        )
 
     async def _send_loop(self, writer: "WriterAsyncIOStream"):
         try:
@@ -658,24 +666,37 @@ class WriterAsyncIOReconnector:
             messages = list(self._messages)
 
             last_seq_no = 0
-            for m in messages:
-                writer.write([m])
+            if messages:
+                writer.write(messages)
+                last_seq_no = messages[-1].seq_no
                 logger.debug(
-                    "writer reconnector %s sent buffered message seqno=%s",
+                    "writer reconnector %s sent %s buffered messages seqno=%s..%s",
                     self._id,
-                    m.seq_no,
+                    len(messages),
+                    messages[0].seq_no,
+                    messages[-1].seq_no,
                 )
-                last_seq_no = m.seq_no
 
             while True:
                 new_msg: InternalMessage = await self._new_messages.get()
-                if new_msg.seq_no > last_seq_no:
-                    writer.write([new_msg])
-                    logger.debug(
-                        "writer reconnector %s sent message seqno=%s",
-                        self._id,
-                        new_msg.seq_no,
-                    )
+                if new_msg.seq_no <= last_seq_no:
+                    continue
+
+                batch = [new_msg]
+                while not self._new_messages.empty():
+                    next_msg = self._new_messages.get_nowait()
+                    if next_msg.seq_no > last_seq_no:
+                        batch.append(next_msg)
+
+                writer.write(batch)
+                last_seq_no = batch[-1].seq_no
+                logger.debug(
+                    "writer reconnector %s sent %s messages seqno=%s..%s",
+                    self._id,
+                    len(batch),
+                    batch[0].seq_no,
+                    batch[-1].seq_no,
+                )
         except asyncio.CancelledError:
             # the loop task cancelled be parent code, for example for reconnection
             # no need to stop all work.